Рефачит таб-скрин и фиксит баг вылета из учетки
