+++ project/src/components/movie-details-screen/movie-details-screen.jsx
@@ -1,17 +1,19 @@
 import React, { useEffect, useState } from 'react';
-import PropTypes from 'prop-types';
+import { connect, useDispatch, useSelector } from 'react-redux';
 import { Link, useParams } from 'react-router-dom';
+import PropTypes from 'prop-types';
 
 import MovieProp from '../../props/movie.prop';
 
 import { AuthorizationStatus, MAX_COUNT_SIMILAR_MOVIES, Tabs } from '../../const';
 
-import MovieDetailsTabs from './movie-details-tabs';
 import MovieListScreen from '../movie-list-screen/movie-list-screen';
 import Logo from '../logo/logo';
 import UserStatus from '../user-status/user-status';
-import { connect, useDispatch, useSelector } from 'react-redux';
 import { fetchSimilarMovies } from '../../store/api-actions';
+import OverviewTab from '../movie-details-screen/overview-tab';
+import DetailsTab from '../movie-details-screen/details-tab';
+import ReviewsTab from '../movie-details-screen/reviews-tab';
 
 export function MovieDetailsScreen(props) {
   const { movies, authorizationStatus } = props;
@@ -26,16 +28,26 @@ export function MovieDetailsScreen(props) {
     year,
   } = movie;
 
+  const getContentFromTab = (tab) => {
+    switch (tab) {
+      case Tabs.OVERVIEW: return <OverviewTab movie={movie} />;
+      case Tabs.DETAILS: return <DetailsTab movie={movie} />;
+      case Tabs.REVIEWS: return <ReviewsTab id={+id} />;
+      default: return 0;
+    }
+  };
+
   const [currentTab, setCurrentTab] = useState(Tabs.OVERVIEW);
-  const dispatch = useDispatch();
   const similarMovies = useSelector((state) => state.similarMovies);
+  const dispatch = useDispatch();
+
   useEffect(() => {
     dispatch(fetchSimilarMovies(id));
-  }, []);
+  }, [id]);
 
-  const handleSetCurrentTab = (tab) => {
-    setCurrentTab(tab);
-  };
+  useEffect(() => {
+    getContentFromTab(currentTab);
+  }, [currentTab]);
 
   return (
     <>
@@ -86,7 +98,18 @@ export function MovieDetailsScreen(props) {
             <div className="film-card__poster film-card__poster--big">
               <img src={movie.poster} alt={`${title} poster`} width="218" height="327" />
             </div>
-            <MovieDetailsTabs currentTab={currentTab} movie={movie} onSetCurrentTab={handleSetCurrentTab} />
+            <div className="film-card__desc">
+              <nav className="film-nav film-card__nav">
+                <ul className="film-nav__list">
+                  {Object.values(Tabs).map((tab) => (
+                    <li key={tab} className={`film-nav__item ${currentTab === tab ? 'film-nav__item--active' : ''}`} >
+                      <a href="#" className="film-nav__link" onClick={(evt) => { evt.preventDefault(); setCurrentTab(tab); }}>{tab}</a>
+                    </li>
+                  ))}
+                </ul>
+              </nav>
+              {getContentFromTab(currentTab)}
+            </div>
           </div>
         </div>
       </section>
