+++ project/src/components/private-route/private-route.jsx
@@ -20,10 +20,10 @@ export function PrivateRoute({ render, path, exact, authorizationStatus }) {
 }
 
 PrivateRoute.propTypes = {
-  authorizationStatus: PropTypes.string.isRequired,
   exact: PropTypes.bool.isRequired,
   path: PropTypes.string.isRequired,
   render: PropTypes.func.isRequired,
+  authorizationStatus: PropTypes.string.isRequired,
 };
 
 const mapStateToProps = (state) => ({
